Remove legacy tally code
SWATCH-687 moved most of the tally logic to the database making the
runtime tally calculations obsolete.  Now that SWATCH-687 has been
deployed and is running successfully, it's time to remove the legacy
code.

See https://issues.redhat.com/browse/SWATCH-892
