+++ src/main/java/org/candlepin/subscriptions/ApplicationProperties.java
@@ -132,9 +132,6 @@ public class ApplicationProperties {
   /** Sets a hard limit on the size of accounts that HBI-based tally will attempt to process. */
   private int tallyMaxHbiAccountSize;
 
-  /** If enabled, nightly tally will calculated via host objects */
-  private boolean legacyNightlyTallyEnabled = false;
-
   /**
    * Interval for system update flush when for reconciliation of HBI data w/ swatch system data.
    *
