+++ src/main/java/org/candlepin/subscriptions/tally/TallySnapshotController.java
@@ -25,7 +25,6 @@ import java.time.OffsetDateTime;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 import org.candlepin.subscriptions.ApplicationProperties;
@@ -110,17 +109,7 @@ public class TallySnapshotController {
 
     AccountUsageCalculation accountCalc;
     try {
-      if (props.isLegacyNightlyTallyEnabled()) {
-        Optional<AccountUsageCalculation> optAccountCalc = performLegacyTally(orgId, account);
-
-        // If no calculation is returned (system threshold reached), abort the tally.
-        if (optAccountCalc.isEmpty()) {
-          return;
-        }
-        accountCalc = optAccountCalc.get();
-      } else {
-        accountCalc = performTally(orgId);
-      }
+      accountCalc = performTally(orgId);
 
       if (props.isCloudigradeEnabled()) {
         attemptCloudigradeEnrichment(accountCalc);
@@ -221,24 +210,6 @@ public class TallySnapshotController {
 
     var calculatedProducts = usageCalculations.getValue().getProducts();
     return calculatedProducts.stream().anyMatch(tagProfile::isProductPAYGEligible);
-  }
-
-  @SuppressWarnings("java:S1874")
-  private Optional<AccountUsageCalculation> performLegacyTally(String orgId, String account) {
-    OrgHostsData orgHostsData =
-        retryTemplate.execute(
-            context -> {
-              try {
-                return usageCollector.collect(this.applicableProducts, account, orgId);
-              } catch (SystemThresholdException e) {
-                log.warn(e.getMessage());
-                return null;
-              }
-            });
-    if (Objects.isNull(orgHostsData)) { // orgHostsData is null when the threshold is breached
-      return Optional.empty();
-    }
-    return Optional.of(usageCollector.tally(this.applicableProducts, orgHostsData));
   }
 
   private AccountUsageCalculation performTally(String orgId) {
