+++ swatch-tally/deploy/clowdapp.yaml
@@ -161,8 +161,6 @@ parameters:
     value: '' # don't restrict to a specific machine pool by default
   - name: JOB_MACHINE_POOL
     value: ''
-  - name: LEGACY_NIGHTLY_TALLY_ENABLED
-    value: 'false'
 
 objects:
 - apiVersion: cloud.redhat.com/v1alpha1
@@ -393,8 +391,6 @@ objects:
               value: ${ENABLE_SYNCHRONOUS_OPERATIONS}
             - name: TALLY_MAX_HBI_ACCOUNT_SIZE
               value: ${TALLY_MAX_HBI_ACCOUNT_SIZE}
-            - name: LEGACY_NIGHTLY_TALLY_ENABLED
-              value: ${LEGACY_NIGHTLY_TALLY_ENABLED}
             - name: HBI_RECONCILIATION_FLUSH_INTERVAL
               value: ${HBI_RECONCILIATION_FLUSH_INTERVAL}
             - name: HOST_LAST_SYNC_THRESHOLD
@@ -904,4 +900,3 @@ objects:
     name: cloudigrade-psk
   data:
     psk: ZHVtbXk=
-
